Added calculation of state completion statistics
This involved keeping track of how many counties there are in each state
in CountyData, which may turn out to be a bad idea.  It's also
duplicated, in that we have that information in the tree view.  Once
again, if we implemented our own model we could eliminate a lot of work.